4-Workbook/HotelApp/src/main/java/com/pluralsight/Room.java
package com.pluralsight;

public class Room {
    private int numberOfBeds;
    private double price;
    private boolean isOccupied;
    private boolean isDirty;
    private boolean isAvailable;

    public void Room(int numberOfBeds, double price, boolean isOccupied, boolean isDirty, boolean isAvailable) {
        this.numberOfBeds = numberOfBeds;
        this.price = price;
        this.isOccupied = isOccupied;
        this.isDirty = isDirty;
        this.isAvailable = isAvailable;
    }

    public int getNumberOfBeds() {
        return numberOfBeds;
    }

    public double getPrice() {
        return price;
    }

    public boolean isOccupied() {
        return isOccupied;
    }

    public boolean isDirty() {
        return isDirty;
    }

    public boolean isAvailable() {
        return isAvailable;
    }
    public void checkIn() {
        isOccupied = true;
        isDirty = true;
        isAvailable = false;
    }
    public void checkOut() {
        isOccupied = false;
        isDirty = true;
        isAvailable = false;
    }
    public void cleanRoom() {
        isDirty = false;
        isAvailable = true;
    }
}
